basic textures

+++ hardcore-sys/hardcore/src/util/reverse_bits.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <array>
-
 #include "number.hpp"
 #include "flow.hpp"
+
+#include <array>
 
 template<class T>
 concept Integral = std::is_integral_v<T>; //!< An integral number type.
@@ -15,21 +15,21 @@ static constexpr Sz reversed_bytes_size = 256; //!< The size of the `reversed_by
  * Indexing into this array with any byte value will yield that same byte with its bits reversed.
  */
 static constexpr std::array<u8, reversed_bytes_size> reversed_bytes{
-	[]() constexpr {
-		auto res = std::array<u8, reversed_bytes_size>();
-		for (Sz i = 0; i < reversed_bytes_size; ++i) {
-			u8 byte = i;
-			u8 count = 8;
-			u8 reversed_byte = 0;
-			while (count) {
-				reversed_byte = (reversed_byte << 1) | (byte & 1);
-				byte = byte >> 1;
-				count--;
-			}
-			res[i] = reversed_byte;
-		}
-		return res;
-	}()
+    []() constexpr {
+        auto res = std::array<u8, reversed_bytes_size>();
+        for (Sz i = 0; i < reversed_bytes_size; ++i) {
+            u8 byte = static_cast<u8>(i);
+            u8 count = 8;
+            u8 reversed_byte = 0;
+            while (count) {
+                reversed_byte = (reversed_byte << 1) | (byte & 1);
+                byte = byte >> 1;
+                count--;
+            }
+            res[i] = reversed_byte;
+        }
+        return res;
+    }()
 };
 
 /**
@@ -42,46 +42,46 @@ static constexpr std::array<u8, reversed_bytes_size> reversed_bytes{
  * @param bytes The array to be reversed.
  */
 template<Sz N>
-constexpr void reverse_byte_array([[maybe_unused]] std::array<u8, N> &bytes) noexcept {
-	static_assert(InstantiatedVarFalse<decltype(N), N>::value, "Impossible number byte size");
+constexpr void reverse_byte_array([[maybe_unused]] std::array<u8, N>& bytes) noexcept {
+    static_assert(InstantiatedVarFalse<decltype(N), N>::value, "Impossible number byte size");
 }
 
 template<>
-[[maybe_unused]] constexpr void reverse_byte_array<1>(std::array<u8, 1> &bytes) noexcept {
-	bytes[0] = reversed_bytes[bytes[0]];
+[[maybe_unused]] constexpr void reverse_byte_array<1>(std::array<u8, 1>& bytes) noexcept {
+    bytes[0] = reversed_bytes[bytes[0]];
 }
 
 template<>
-[[maybe_unused]] constexpr void reverse_byte_array<2>(std::array<u8, 2> &bytes) noexcept {
-	const u8 tmp = reversed_bytes[bytes[0]];
-	bytes[0] = reversed_bytes[bytes[1]];
-	bytes[1] = tmp;
+[[maybe_unused]] constexpr void reverse_byte_array<2>(std::array<u8, 2>& bytes) noexcept {
+    const u8 tmp = reversed_bytes[bytes[0]];
+    bytes[0] = reversed_bytes[bytes[1]];
+    bytes[1] = tmp;
 }
 
 template<>
-[[maybe_unused]] constexpr void reverse_byte_array<4>(std::array<u8, 4> &bytes) noexcept {
-	u8 tmp = reversed_bytes[bytes[0]];
-	bytes[0] = reversed_bytes[bytes[3]];
-	bytes[3] = tmp;
-	tmp = reversed_bytes[bytes[1]];
-	bytes[1] = reversed_bytes[bytes[2]];
-	bytes[2] = tmp;
+[[maybe_unused]] constexpr void reverse_byte_array<4>(std::array<u8, 4>& bytes) noexcept {
+    u8 tmp = reversed_bytes[bytes[0]];
+    bytes[0] = reversed_bytes[bytes[3]];
+    bytes[3] = tmp;
+    tmp = reversed_bytes[bytes[1]];
+    bytes[1] = reversed_bytes[bytes[2]];
+    bytes[2] = tmp;
 }
 
 template<>
-[[maybe_unused]] constexpr void reverse_byte_array<8>(std::array<u8, 8> &bytes) noexcept {
-	u8 tmp = reversed_bytes[bytes[0]];
-	bytes[0] = reversed_bytes[bytes[7]];
-	bytes[7] = tmp;
-	tmp = reversed_bytes[bytes[1]];
-	bytes[1] = reversed_bytes[bytes[6]];
-	bytes[6] = tmp;
-	tmp = reversed_bytes[bytes[2]];
-	bytes[2] = reversed_bytes[bytes[5]];
-	bytes[5] = tmp;
-	tmp = reversed_bytes[bytes[3]];
-	bytes[3] = reversed_bytes[bytes[4]];
-	bytes[4] = tmp;
+[[maybe_unused]] constexpr void reverse_byte_array<8>(std::array<u8, 8>& bytes) noexcept {
+    u8 tmp = reversed_bytes[bytes[0]];
+    bytes[0] = reversed_bytes[bytes[7]];
+    bytes[7] = tmp;
+    tmp = reversed_bytes[bytes[1]];
+    bytes[1] = reversed_bytes[bytes[6]];
+    bytes[6] = tmp;
+    tmp = reversed_bytes[bytes[2]];
+    bytes[2] = reversed_bytes[bytes[5]];
+    bytes[5] = tmp;
+    tmp = reversed_bytes[bytes[3]];
+    bytes[3] = reversed_bytes[bytes[4]];
+    bytes[4] = tmp;
 }
 
 /**
@@ -93,7 +93,7 @@ template<>
  */
 template<Integral T>
 constexpr T reverse_bits(T number) noexcept {
-	auto bytes = std::bit_cast<std::array<u8, sizeof(T)> >(number);
-	reverse_byte_array(bytes);
-	return std::bit_cast<T>(bytes);
+    auto bytes = std::bit_cast<std::array<u8, sizeof(T)>>(number);
+    reverse_byte_array(bytes);
+    return std::bit_cast<T>(bytes);
 }
